Adaptation of parser to run client-side
I've modified the parser so it runs on the client side. It will simply return a string with the entire program. So pretty much what it was writing to console.

Only drawback is that it doesn't have `include` support. You can fix that with ajax magic, of course. I didn't bother to.

Great work, keep it up.